feat: export goBack ts intetface
