+++ src/components/goBack/goBack.tsx
@@ -1,13 +1,8 @@
 import React from 'react'
 import { Icon } from 'antd/lib'
 import { browserHistory, hashHistory } from 'react-router'
+import { GoBackProps } from './index'
 
-export interface GoBackProps {
-    url?: string;
-    autoClose?: boolean;
-    style?: React.CSSProperties;
-    history?: boolean;
-}
 export default class GoBack extends React.Component<GoBackProps, any> {
     go = () => {
         const { url, history, autoClose } = this.props
