+++ src/components/goBack/index.tsx
@@ -1,5 +1,15 @@
-import GoBack from './goBack'
-import GoBackButton from './goBackButton'
+import GoBack from './goBack';
+import GoBackButton from './goBackButton';
 
-GoBack.GoBackButton = GoBackButton
-export default GoBack
+export interface GoBackProps {
+    url?: string;
+    autoClose?: boolean;
+    style?: React.CSSProperties;
+    history?: boolean;
+}
+export interface GoBackButtonProps extends GoBackProps{
+    title?: string;
+}
+
+GoBack.GoBackButton = GoBackButton;
+export default GoBack;
